+++ src/Exception/InvalidArgumentException.php
@@ -3,6 +3,6 @@ declare(strict_types=1);
 
 namespace NumPHP\Exception;
 
-class IllegalArgumentException extends \Exception
+class InvalidArgumentException extends \InvalidArgumentException
 {
 }